ManageACI/requirements.txt
azure-common==1.1.25
azure-mgmt-containerinstance==2.0.0
PyYAML==5.3.1